requirements-dev.txt
# 開発用依存関係を追加
streamlit
gspread
google-auth
beautifulsoup4
requests

# 開発・テスト用パッケージ
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950

# 型チェック用
types-requests
